# gateway/docs/swagger.yaml
definitions:
...
  genproto.AddReviewReq:
    properties:
      booking_id:
        type: string
      comment:
        type: string
      provider_id:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
  genproto.AddServiceReq:
    properties:
      description:
        type: string
      duration:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  genproto.Booking:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      provider_id:
        $ref: '#/definitions/genproto.Provider'
      schudule_time:
        type: string
      service_id:
        $ref: '#/definitions/genproto.Services'
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  genproto.GetServicesByPriceRangeResp:
    properties:
      services:
        items:
          $ref: '#/definitions/genproto.Services'
        type: array
    type: object
  genproto.GetServicesResp:
    properties:
      services:
        $ref: '#/definitions/genproto.Services'
    type: object
...
  genproto.ListAllReviewsResp:
    properties:
      reviews:
        items:
          $ref: '#/definitions/genproto.Review'
        type: array
    type: object
  genproto.ListAllServicesResp:
    properties:
      services:
        items:
          $ref: '#/definitions/genproto.Services'
        type: array
    type: object
...
  genproto.Provider:
    properties:
      availability:
        type: string
      average_rating:
        type: number
      company_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      services:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.Review:
    properties:
      booking_id:
        $ref: '#/definitions/genproto.Booking'
      comment:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      provider_id:
        $ref: '#/definitions/genproto.Provider'
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.SearchServicessResp:
    properties:
      services:
        items:
          $ref: '#/definitions/genproto.Services'
        type: array
    type: object
  genproto.Services:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  genproto.UpdateServiceReq:
    properties:
      id:
        type: string
      services:
        $ref: '#/definitions/genproto.Services'
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/review:
    get:
      consumes:
      - application/json
      description: This API Lists a new review
      parameters:
      - description: Booking iD
        in: query
        name: booking_id
        type: string
      - description: provider ID
        in: query
        name: provider_id
        type: string
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ListAllReviewsResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all review
      tags:
      - carwash/review
  /api/v1/review/{id}:
    delete:
      consumes:
      - application/json
      description: This API deleted a new review
      parameters:
      - description: review Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: review deleted successfully
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete review
      tags:
      - carwash/review
    get:
      consumes:
      - application/json
      description: This API Gets a  review
      parameters:
      - description: review Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Review'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get review
      tags:
      - carwash/review
  /api/v1/review/add:
    post:
      consumes:
      - application/json
      description: This API creates a review
      parameters:
      - description: review
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.AddReviewReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "created successfully'
          schema:
            type: string
        "400":
          description: 'error": "error message'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: CREATE review
      tags:
      - carwash/review
  /api/v1/service:
    get:
      consumes:
      - application/json
      description: This API Lists a new service
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Description
        in: query
        name: description
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ListAllServicesResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all service
      tags:
      - carwash/service
  /api/v1/service/{id}:
    delete:
      consumes:
      - application/json
      description: This API deleted a new service
      parameters:
      - description: service Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: service deleted successfully
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete service
      tags:
      - carwash/service
    get:
      consumes:
      - application/json
      description: This API Gets a  service
      parameters:
      - description: service Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.GetServicesResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get service
      tags:
      - carwash/service
    put:
      consumes:
      - application/json
      description: This API Put s a new service
      parameters:
      - description: Put  service Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: service updated successfully
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Put  service
      tags:
      - carwash/service
  /api/v1/service/add:
    post:
      consumes:
      - application/json
      description: This API creates a service
      parameters:
      - description: service
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.AddServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "created successfully'
          schema:
            type: string
        "400":
          description: 'error": "error message'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: CREATE service
      tags:
      - carwash/service
  /api/v1/service/priceRange:
    get:
      consumes:
      - application/json
      description: This API gets a  service
      parameters:
      - description: min_price
        in: path
        name: min_price
        required: true
        type: integer
      - description: max_price
        in: path
        name: max_price
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.GetServicesByPriceRangeResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get service
      tags:
      - carwash/service
  /api/v1/service/search:
    get:
      consumes:
      - application/json
      description: This API gets a  service
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: description
        in: path
        name: description
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.SearchServicessResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get service
      tags:
      - carwash/service
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"